separate emulation_nav from emulation_base
